requirements.txt
numpy==2.2.6
pandas==2.2.3
psycopg==3.2.9
psycopg-binary==3.2.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
typing_extensions==4.13.2
tzdata==2025.2
